moved notification command into the notifications module
