+++ src/masonite/commands/__init__.py
@@ -10,5 +10,4 @@ from .QueueFailedCommand import QueueFailedCommand
 from .MakeControllerCommand import MakeControllerCommand
 from .MakeJobCommand import MakeJobCommand
 from .MakeMailableCommand import MakeMailableCommand
-from .MakeNotificationCommand import MakeNotificationCommand
-from .NotificationTableCommand import NotificationTableCommand
+
