+++ src/masonite/notification/commands/__init__.py
@@ -1,0 +1,2 @@
+from .MakeNotificationCommand import MakeNotificationCommand
+from .NotificationTableCommand import NotificationTableCommand
