+++ src/masonite/notification/providers/__init__.py
@@ -1,0 +1,1 @@
+from .NotificationProvider import NotificationProvider
