+++ src/masonite/providers/__init__.py
@@ -9,7 +9,6 @@ from .MailProvider import MailProvider
 from .SessionProvider import SessionProvider
 from .HelpersProvider import HelpersProvider
 from .QueueProvider import QueueProvider
-from .NotificationProvider import NotificationProvider
 from .CacheProvider import CacheProvider
 from ..events.providers import EventProvider
 from ..filesystem.providers import StorageProvider
